move GroovyJob and JobState to package scheduler.job
