+++ celements-scheduler/src/main/java/com/celements/scheduler/SchedulerPluginApi.java
@@ -26,6 +26,8 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.quartz.SchedulerException;
 
+import com.celements.scheduler.job.GroovyJob;
+import com.celements.scheduler.job.JobState;
 import com.xpn.xwiki.XWikiContext;
 import com.xpn.xwiki.XWikiException;
 import com.xpn.xwiki.api.Document;
