+++ celements-scheduler/src/main/java/com/celements/scheduler/job/GroovyJob.java
@@ -17,7 +17,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package com.celements.scheduler;
+package com.celements.scheduler.job;
 
 import org.codehaus.groovy.control.CompilationFailedException;
 import org.quartz.Job;
@@ -27,6 +27,7 @@ import org.quartz.JobExecutionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.celements.scheduler.AbstractJob;
 import com.xpn.xwiki.XWikiContext;
 import com.xpn.xwiki.doc.XWikiDocument;
 import com.xpn.xwiki.objects.BaseObject;
