+++ celements-scheduler/src/main/java/com/celements/scheduler/job/JobState.java
@@ -17,7 +17,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package com.celements.scheduler;
+package com.celements.scheduler.job;
 
 import org.quartz.Trigger;
